*refactoring of ConvertFloorNode.
Move ConvertFloorNode from OnnxRuntime.h into a dedicated ConvertFloorNode.cpp
and register it in the CvtOnnx VS2022 project files.
