+++ prj/vs2022/CvtOnnx.vcxproj
@@ -51,6 +51,7 @@
     <ClCompile Include="..\..\src\Cvt\OnnxRuntime\ConvertExpNode.cpp" />
     <ClCompile Include="..\..\src\Cvt\OnnxRuntime\ConvertExpandNode.cpp" />
     <ClCompile Include="..\..\src\Cvt\OnnxRuntime\ConvertFlattenNode.cpp" />
+    <ClCompile Include="..\..\src\Cvt\OnnxRuntime\ConvertFloorNode.cpp" />
     <ClCompile Include="..\..\src\Cvt\OnnxRuntime\ConvertGatherNode.cpp" />
     <ClCompile Include="..\..\src\Cvt\OnnxRuntime\ConvertGemmNode.cpp" />
     <ClCompile Include="..\..\src\Cvt\OnnxRuntime\ConvertGlobalAveragePoolNode.cpp" />
